fix(clear): make re-use type
