+++ src/user-event/utils.js
@@ -1,0 +1,5 @@
+function isInputElement(element) {
+  return element.tagName.toLowerCase() === 'input'
+}
+
+export {isInputElement}
